jobs/pyproject.toml
[project]
name = "asx-trading-lab-jobs"
version = "0.1.0"
description = "ASX Trading Lab - Data ingestion and analytics jobs runner"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "supabase>=2.0.0",
    "yfinance>=0.2.40",
    "pandas>=2.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pandas-stubs>=2.0.0",
]

[project.scripts]
asx-jobs = "asx_jobs.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/asx_jobs"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
